Added route to edit blogpost/id within the blogPostRoutes. Fixed issue where when user created a new blogpsot the user was taken to the homepage by redirecting the user to their specific blogpost they made. Updated and added update route for the user to updated an individual post. Edited syntax in the dashboard.js file. Added editBlogPost.js file that will soley handle the editing blogpost functionality. Seperated the edit blogpost handler from the dashboard handler to avoid conflicts with js files being in the handlebars. Added editBlogPost handlebars that is rendered when the user clicks on the edit button within their blogpost from their dashboard. The route IS working, and the blog can be edited, but there are errors within the edit button not rendering. Also added some features, such as a confirmation before deleting their blog, and the a message indicating the user can be the first to comment under a blog if there are no comments there.

+++ public/js/comments.js
@@ -1,10 +1,8 @@
 const newCommentHandler = async (event) => {
     event.preventDefault();
 
-    // Use the name attribute to select the input field
     const blogPostId = event.target.querySelector('#blogPost_id').value;
-    const content = event.target.querySelector('#userComment').value.trim();
-
+    const content = event.target.querySelector('#userComment').value
 
 
     if (content) {
@@ -15,10 +13,9 @@ const newCommentHandler = async (event) => {
                 'Content-Type': 'application/json',
             },
         });
-        
+
         if (response.ok) {
-            //checking to make sure comments are created
-            console.log(response)
+
             document.location.reload();
         } else {
             alert('There was an problem submitting your comment. Please try again.');
@@ -29,20 +26,25 @@ const newCommentHandler = async (event) => {
 
 const delCommentHandler = async (event) => {
     if (event.target.matches('.delete-comment-button')) {
-        const id = event.target.getAttribute('data-id');
-        
-        const response = await fetch(`/comments/delete/${id}`, {
-            method: 'DELETE',
-        });
-        
-        if (response.ok) {
-            document.location.reload();
-        } else {
-            alert('Failed to delete comment');
+        const confirmed = confirm("Are you sure you want to delete this comment?")
+
+        if (confirmed) {
+
+            const id = event.target.getAttribute('data-id');
+
+            const response = await fetch(`/comments/delete/${id}`, {
+                method: 'DELETE',
+            });
+
+            if (response.ok) {
+                document.location.reload();
+            } else {
+                alert('Failed to delete comment');
+            }
         }
+        // console.log(delCommentHandler)
     }
-    console.log(delCommentHandler)
-};
+}
 
 document.addEventListener('click', delCommentHandler);
